+++ modules/effects/src/passes/postprocessing/image-fun-filters/edgework.ts
@@ -11,7 +11,7 @@ struct edgeWorkUniforms {
   mode: u32,
 };
 
-@group(0) @binding(1) var<uniform> edgeWork: edgeWorkUniforms;
+@group(0) @binding(auto) var<uniform> edgeWork: edgeWorkUniforms;
 
 fn edgeWork_sampleColorRGB(
   sourceTexture: texture_2d<f32>,