+++ modules/effects/src/passes/postprocessing/image-warp-filters/bulgepinch.ts
@@ -12,7 +12,7 @@ struct bulgePinchUniforms {
   center: vec2f,
 };
 
-@group(0) @binding(1) var<uniform> bulgePinch: bulgePinchUniforms;
+@group(0) @binding(auto) var<uniform> bulgePinch: bulgePinchUniforms;
 
 fn bulgePinch_warp(coordIn: vec2f, texCenter: vec2f) -> vec2f {
   var coord = coordIn - texCenter;