+++ modules/gltf/test/index.ts
@@ -8,5 +8,6 @@ import './gltf/gltf.spec';
 import './gltf/gltf-extension-support.spec';
 import './gltf/lights.spec';
 import './parsers/parse-gltf-animations.spec';
+import './parsers/parse-gltf.spec';
 import './parsers/parse-pbr-compressed-texture.spec';
 import './parsers/parse-pbr-material.spec';